rename SpecificMethodBoolNullRector to AssertSameBoolNullToSpecificMethodRector
