+++ tests/Rector/Contrib/PHPUnit/SpecificMethod/AssertSameBoolNullToSpecificMethodRector/Wrong/wrong.php.inc
