--- tests/Rector/Contrib/PHPUnit/SpecificMethodBoolNullRector/Test.php
@@ -1,25 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Tests\Rector\Contrib\PHPUnit\SpecificMethodBoolNullRector;
-
-use Rector\Rector\Contrib\PHPUnit\SpecificMethodBoolNullRector;
-use Rector\Testing\PHPUnit\AbstractRectorTestCase;
-
-final class Test extends AbstractRectorTestCase
-{
-    public function test(): void
-    {
-        $this->doTestFileMatchesExpectedContent(
-            __DIR__ . '/Wrong/wrong.php.inc',
-            __DIR__ . '/Correct/correct.php.inc'
-        );
-    }
-
-    /**
-     * @return string[]
-     */
-    protected function getRectorClasses(): array
-    {
-        return [SpecificMethodBoolNullRector::class];
-    }
-}
